Rework Error message and Dummy customer. Added Api controllers

+++ DummyDataGenerators.DTO/DummyErrorDTO/DummyError.cs
@@ -14,7 +14,7 @@ namespace DummyDataGenerators.DTO.DummyErrorDTO
         public string OS { get; set; }
         public Severity ErrorSeverity { get; set; }
         public string ErrorCode { get; set; }
-        public byte[] IP { get; set; }
+        public int[] IP { get; set; }
         public string Comment { get; set; }
     }
 }
